refactor CPU_Unpickler
